Revert "Add 2FA to login process"
