+++ .env.example
@@ -1,5 +1,2 @@
 APPLICANTS_APP_HOST=bops-care.link:3001
 OS_VECTOR_TILES_API_KEY=xxxxx
-2FA_ENABLED=true
-OTP_SECRET_ENCRYPTION_KEY=xxxxx
-NOTIFY_API_KEY=xxxxx
