+++ Gemfile
@@ -14,7 +14,6 @@ gem "business_time"
 gem "daemons"
 gem "delayed_job_active_record"
 gem "devise"
-gem "devise-two-factor"
 gem "faker", require: false
 gem "faraday", require: false
 gem "govuk_design_system_formbuilder"
@@ -24,7 +23,6 @@ gem "image_processing", "~> 1.12"
 gem "jbuilder"
 gem "mail-notify"
 gem "mini_magick"
-gem "notifications-ruby-client"
 gem "pg", ">= 0.18", "< 2.0"
 gem "puma", "~> 4"
 gem "rails", "~> 6.1.4.7"
