--- app/views/devise/sessions/setup.html.erb
@@ -1,15 +1,0 @@
-<div class="govuk-grid-row">
-  <div class="govuk-grid-column-two-thirds">
-    <h1 class="govuk-heading-l">
-        Enter your phone number
-    </h1>
-    <p class="govuk-body"> To use two-factor authentication we need a UK mobile phone number for your account.</p>
-    <div class=" govuk-form-group">
-      <%= form_for(@user, as: :user, url: user_session_path, method: :post) do |form| %>
-        <%= form.label :mobile_number, class: "govuk-label" %>
-        <%= form.text_field :mobile_number, class: "govuk-textarea govuk-input" %>
-        <%= form.submit "Send code", class: "govuk-button", data: { module: "govuk-button" } %>
-      <% end %>
-    </div>
-  </div>
-</div>
