--- app/views/devise/sessions/two_factor.html.erb
@@ -1,26 +1,0 @@
-<div class="govuk-grid-row">
-  <div class="govuk-grid-column-two-thirds">
-    <h1 class="govuk-heading-l">
-      Enter the code you have recieved by text message
-    </h1>
-
-    <p class="govuk-body">A 6-digit code has been sent to your mobile phone. This message may take a minute to arrive.</p>
-
-    <div class=" govuk-form-group ">
-      <%= form_for(@user, as: :user, url: user_session_path, method: :post) do |form| %>
-        <%= form.label :otp_attempt, class: "govuk-label" %>
-        <%= form.text_field :otp_attempt, class: "govuk-textarea govuk-input" %>
-
-        <%= link_to "Resend code", resend_code_path, class: "govuk-link" %>
-
-        <p class="govuk-phase-banner__content govuk-!-padding-top-3 govuk-!-padding-bottom-3">
-          <span class="govuk-phase-banner__text">
-            If you have an issue logging in, please send an email to <%= mail_to @user.local_authority.email_address, @user.local_authority.email_address, class: "govuk-link" %>
-          </span>
-        </p>
-
-        <%= form.submit "Enter code", class: "govuk-button", data: { module: "govuk-button" } %>
-      <% end %>
-    </div>
-  </div>
-</div>
