+++ config/initializers/devise.rb
@@ -1,14 +1,6 @@
 # frozen_string_literal: true
 
 Devise.setup do |config|
-  config.warden do |manager|
-    manager.default_strategies(scope: :user).unshift :two_factor_authenticatable
-  end
-
-  config.sign_in_after_reset_password = false
-
-  config.timeout_in = 6.hours
-
   config.mailer_sender = "support@unboxed.co"
 
   config.mailer = "DeviseMailer"
