+++ config/initializers/filter_parameter_logging.rb
@@ -4,5 +4,5 @@
 
 # Configure sensitive parameters which will be filtered from the log file.
 Rails.application.config.filter_parameters += %i[
-  passw secret token _key crypt salt certificate otp ssn otp_attempt
+  passw secret token _key crypt salt certificate otp ssn
 ]
