--- db/migrate/20220408091235_add_mobile_number_to_users.rb
@@ -1,7 +1,0 @@
-# frozen_string_literal: true
-
-class AddMobileNumberToUsers < ActiveRecord::Migration[6.1]
-  def change
-    add_column :users, :mobile_number, :string, unique: true
-  end
-end
